DB Schema Update

+++ database/migrations/2020_07_14_172105_create_direct_chats_table.php
@@ -14,7 +14,7 @@ class CreateDirectChatsTable extends Migration
     public function up()
     {
         Schema::create('direct_chats', function (Blueprint $table) {
-            $table->unsignedInteger('id')->index();
+            $table->unsignedBigInteger('id')->index();
             $table->timestamps();
         });
     }
