+++ database/migrations/2020_08_08_062133_create_messages_table.php
@@ -15,7 +15,7 @@ class CreateMessagesTable extends Migration
     {
         Schema::create('messages', function (Blueprint $table) {
             $table->unsignedInteger('id')->index();
-            $table->unsignedInteger('chat_id');
+            $table->unsignedBigInteger('chat_id');
             $table->foreignId('from_id')
                 ->constrained('users')
                 ->onDelete('cascade');
