+++ database/migrations/2020_08_08_063810_create_chat_members_table.php
@@ -15,7 +15,7 @@ class CreateChatMembersTable extends Migration
     {
         Schema::create('chat_members', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('chat_id');
+            $table->unsignedBigInteger('chat_id');
             $table->foreignId('user_id')
                 ->constrained('users')
                 ->onDelete('cascade');
